+++ packages/blade/src/components/SpotlightPopoverTour/TourFooter.native.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable react/jsx-no-useless-fragment */
+/* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */
+import { throwBladeError } from '~utils/logger';
+
+const SpotlightPopoverTourFooter = (): React.ReactElement => {
+  throwBladeError({
+    message: 'Tour is not yet implemented for native',
+    moduleName: 'Tour',
+  });
+
+  return <></>;
+};
+
+export { SpotlightPopoverTourFooter };