+++ packages/blade/src/utils/storybook/componentStatusData.ts
@@ -326,12 +326,12 @@ const componentData: ComponentStatusDataType = [
   },
   {
     name: 'FileUpload',
-    status: 'planned-Q2-dev',
+    status: 'to-be-decided',
     description: 'FileUpload component allow users to select one or more files to upload.',
   },
   {
     name: 'Drawer',
-    status: 'planned-Q2-design',
+    status: 'to-be-decided',
     description: 'The Drawer component is a panel that slides out from the edge of the screen.',
   },
   {
@@ -359,14 +359,18 @@ const componentData: ComponentStatusDataType = [
   },
   {
     name: 'Tabs',
-    status: 'in-design',
+    status: 'released',
+    releasedIn: '10.16.0',
+    storybookLink: 'Components/Tabs',
     description:
       'Tabs is a component which will allow you to show multiple clickable tabs in your UI',
   },
   {
-    name: 'Data Table',
-    status: 'in-design',
-    description: 'DataTable will allow you to display your data in tabular manner',
+    name: 'Table',
+    status: 'released',
+    releasedIn: '10.19.0',
+    storybookLink: 'Components/Table',
+    description: 'Table will allow you to display your data in tabular manner',
   },
   {
     name: 'Date Picker',
@@ -383,12 +387,12 @@ const componentData: ComponentStatusDataType = [
   },
   {
     name: 'Toast',
-    status: 'planned-Q2-dev',
+    status: 'to-be-decided',
     description: 'Toast is a component to show a simple floating messages to your users',
   },
   {
     name: 'Button Group',
-    status: 'planned-Q2-dev',
+    status: 'to-be-decided',
     description: 'ButtonGroup component can be used to group related buttons.',
   },
   {